conditioned.rb resource + fenestration PSI tests
