+++ lib/measures/tbd/README.md
@@ -2,13 +2,13 @@
 
 ###### (Automatically generated documentation)
 
-# Thermal Bridging and Derating - TBD
+# 
 
 ## Description
-Thermally derates opaque constructions from major thermal bridges
+
 
 ## Modeler Description
-(see github.com/rd2/tbd)
+
 
 ## Measure Type
 ModelMeasure
@@ -19,46 +19,8 @@ ModelMeasure
 ## Arguments
 
 
-### Load tbd.json
-Loads existing tbd.json from model files directory, overrides other arguments if true.
-**Name:** load_tbd_json,
-**Type:** Boolean,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
-
-### Default thermal bridge option to use if not reading tbd.json
-e.g. poor, regular, efficient, code
-**Name:** option,
-**Type:** Choice,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
-
-### Write tbd.out.json
-Write tbd.out.json to customize for subsequent runs. Edit and place in model files directory as tbd.json
-**Name:** write_tbd_json,
-**Type:** Boolean,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
-
-### Generate Kiva inputs
-Generate OSM Kiva settings and objects if model surfaces have 'foundation' boundary conditions
-**Name:** gen_kiva,
-**Type:** Boolean,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
-
-### Force-generate Kiva inputs
-Overwrites all 'ground' boundary conditions as 'foundation' before generating OSM Kiva inputs
-**Name:** gen_kiva_force,
-**Type:** Boolean,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
 
 
+This measure does not have any user arguments
